trader validation
